Extract mobile foundation helpers
