+++ apps/field-mobile/src/components/LabeledInput.tsx
@@ -1,0 +1,59 @@
+import React from "react";
+import { StyleSheet, Text, TextInput, View } from "react-native";
+
+export type LabeledInputProps = {
+  label: string;
+  value: string;
+  onChangeText: (text: string) => void;
+  keyboardType?: "default" | "number-pad" | "decimal-pad";
+  multiline?: boolean;
+  secureTextEntry?: boolean;
+};
+
+export function LabeledInput({
+  label,
+  value,
+  onChangeText,
+  keyboardType = "default",
+  multiline = false,
+  secureTextEntry = false,
+}: LabeledInputProps) {
+  return (
+    <View style={styles.inputWrap}>
+      <Text style={styles.label}>{label}</Text>
+      <TextInput
+        style={[styles.input, multiline && styles.multilineInput]}
+        value={value}
+        onChangeText={onChangeText}
+        keyboardType={keyboardType}
+        multiline={multiline}
+        secureTextEntry={secureTextEntry}
+      />
+    </View>
+  );
+}
+
+const styles = StyleSheet.create({
+  inputWrap: {
+    marginBottom: 8,
+  },
+  label: {
+    marginBottom: 4,
+    fontSize: 13,
+    color: "#334155",
+  },
+  input: {
+    borderWidth: 1,
+    borderColor: "#cbd5e1",
+    borderRadius: 8,
+    backgroundColor: "#ffffff",
+    paddingHorizontal: 10,
+    paddingVertical: 8,
+    fontSize: 14,
+    color: "#111827",
+  },
+  multilineInput: {
+    minHeight: 70,
+    textAlignVertical: "top",
+  },
+});
